Fixed bug which prevented usage from gulp.
This patch fixes that, by removing the use of pn to open the file,
which has already been read into memory by gulp anyway.

svg2png already uses promises so there is no effect on synchronousity.

Added an extra error message to remind users that only Buffers are
supported, not streams. I'm not sure if this will affect anybody.

An added bonus is that any changes made by gulp to the file will
now be reflected by gulp-svg2png.